ui: Move `IconDecoration` and `DecoratedIcon` to their own modules
